# swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Agrirouter Middleware",
    "description": "Middleware to access the agrirouter using a 'smaller' API. The middleware will handle the connection for all endpoints and fetches / stores messages from the agrirouter within a custom database.",
    "contact": {
      "name": "Agrirouter Middleware",
      "url": "https://github.com/agrirouter-middleware"
    },
    "license": {
      "name": "Apache License Version 2.0",
      "url": "https://github.com/agrirouter-middleware/LICENSE"
    },
    "version": "1.0.0"
  },
...
  "paths": {
    "/unsecured/api/maintenance/endpoint/resend/capabilities-and-subscriptions/{internalEndpointId}": {
      "put": {
        "tags": [
          "maintenance"
        ],
        "description": "Resend the capabilities and subscriptions for the endpoint.",
        "operationId": "maintenance.resend-capabilities",
        "parameters": [
          {
            "name": "internalEndpointId",
            "in": "path",
            "description": "The internal endpoint ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "In case the operation was successful.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/unsecured/api/maintenance/endpoint/disable/{internalEndpointId}": {
      "put": {
        "tags": [
          "maintenance"
        ],
        "description": "Disable a specific endoint.",
        "operationId": "maintenance.disable",
        "parameters": [
          {
            "name": "internalEndpointId",
            "in": "path",
            "description": "The internal endpoint ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "In case the operation was successful.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/unsecured/api/callback": {
      "get": {
        "tags": [
          "agrirouter callback"
        ],
        "description": "The callback for the onboard process. Used by the agrirouter to send the onboard process data.",
        "operationId": "callback.callback",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "description": "The state to identify the request internally.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "description": "Token containing the information for the onboard process, generated by the agrirouter.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "signature",
            "in": "query",
            "description": "Signature from the agrirouter, used for validation.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "description": "Error information, in case the onboard process failed.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "In case of a successful response, the callback redirects to an internal page to show the result of the onboard process.",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/secured/api/application/{applicationId}/endpoints": {
      "get": {
        "tags": [
          "application management"
        ],
        "description": "Fetch all endpoints for an application.",
        "operationId": "application.endpoints",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "description": "The ID of the application.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The endpoints for the application.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindEndpointsForApplicationResponse"
                }
              }
            }
          },
          "400": {
            "description": "In case of a business exception.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "In case of an unknown error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "secured-api": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "The timestamp (UTC) of the response.",
            "format": "date-time"
          },
          "key": {
            "type": "string",
            "description": "The key of the error."
          },
          "errorMessage": {
            "type": "string",
            "description": "The message of the error."
          }
        },
        "description": "Generic error response to map the internal errors and provide clean API description."
      },
...
      "EndpointDto": {
        "type": "object",
        "properties": {
          "agrirouterEndpointId": {
            "type": "string",
            "description": "The ID of an endpoint."
          },
          "externalEndpointId": {
            "type": "string",
            "description": "The external ID of an endpoint."
          },
          "endpointType": {
            "type": "string",
            "description": "The type of the endpoint.",
            "enum": [
              "NON_VIRTUAL",
              "VIRTUAL"
            ]
          },
          "deactivated": {
            "type": "boolean",
            "description": "Marks an endpoint as deactivated."
          },
          "agrirouterAccountId": {
            "type": "string",
            "description": "The account ID for this endpoint."
          }
        },
        "description": "Representation of an endpoint."
      },
      "EndpointWithChildrenDto": {
        "type": "object",
        "properties": {
          "agrirouterEndpointId": {
            "type": "string",
            "description": "The ID of an endpoint."
          },
          "externalEndpointId": {
            "type": "string",
            "description": "The external ID of an endpoint."
          },
          "endpointType": {
            "type": "string",
            "description": "The type of the endpoint.",
            "enum": [
              "NON_VIRTUAL",
              "VIRTUAL"
            ]
          },
          "connectedVirtualEndpoints": {
            "type": "array",
            "description": "All the children of the endpoint.",
            "items": {
              "$ref": "#/components/schemas/EndpointDto"
            }
          },
          "deactivated": {
            "type": "boolean",
            "description": "Marks an endpoint as deactivated."
          },
          "agrirouterAccountId": {
            "type": "string",
            "description": "The account ID for this endpoint."
          }
        },
        "description": "Representation of an endpoint with all children."
      },
      "FindEndpointsForApplicationResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "The timestamp (UTC) of the response.",
            "format": "date-time"
          },
          "endpoints": {
            "type": "array",
            "description": "The endpoints for this application.",
            "items": {
              "$ref": "#/components/schemas/EndpointWithChildrenDto"
            }
          }
        },
        "description": "The response when searching for endpoints belonging to the application."
      },
...
    },
    "securitySchemes": {
      "secured": {
        "type": "http",
        "description": "Default security scheme, used for the private flows.",
        "scheme": "basic"
      }
    }
  }
}